+++ src/Facades/Mark.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace LaraZeus\Mark\Facades;
+
+use Illuminate\Support\Facades\Facade;
+
+class Mark extends Facade
+{
+    protected static function getFacadeAccessor()
+    {
+        return \LaraZeus\Mark\Mark::class;
+    }
+}
