+++ stubs/mark.migration.create.stub
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('user_id')->constrained()->cascadeOnUpdate();
             $table->morphs('markable');
-            $table->string('value')->nullable();
+            $table->string('value');
             $table->json('metadata')->nullable();
             $table->timestamps();
             $table->unique(['user_id', 'markable_type', 'markable_id']);
